refactor(cli): split combined.rs into orchestrator, output, orientation, impact modules
combined.rs concentrated analysis orchestration, the output-format printers,
the orientation header, entry-point display, and impact recording in a single
large file, making it the repo's merge-conflict magnet.

Split it into a combined/ module: mod.rs keeps the orchestrator
(CombinedOptions, resolve_analyses, run_combined, run_combined_dupes,
build_health_opts); output.rs holds the format printers; orientation.rs the
header and entry-point display; impact.rs the telemetry recording. Pure code
motion with zero behavior change (visibility and use-paths only); the
rayon::join parallelism and shared-parse threading are untouched. Test and
snapshot suites pass with no modifications.
